+++ browser/hy/index.test.ts
@@ -1,7 +1,7 @@
 import { JSDOM } from 'jsdom'
 import { test } from 'uvu'
 import { equal, throws } from 'uvu/assert'
-import { attach, hy__bind, tags } from '../index.js'
+import { attach, hy_op, tags } from '../index.js'
 let jsdom:JSDOM
 let prev__window:Window
 let prev__document:Document
@@ -27,23 +27,23 @@ test.after(()=>{
 	globalThis.Text = prev__Text
 	globalThis.Node = prev__Node
 })
-test('hy__bind', with_connected_dom(async connected_dom=>{
+test('hy_op', with_connected_dom(async connected_dom=>{
 	const el_a:Element[] = []
 	const fn0 = (el:Element)=>el_a.push(el)
-	const div0 = div({ 'hy__bind': 'fn0' })
+	const div0 = div({ 'hy_op': 'fn0' })
 	const fn1 = (el:Element)=>el_a.push(el)
-	const div1 = div({ 'hy__bind': 'fn1' })
+	const div1 = div({ 'hy_op': 'fn1' })
 	attach(connected_dom, div(), div0, div1, div())
-	hy__bind(document, { fn0, fn1 })
+	hy_op(document, { fn0, fn1 })
 	equal(el_a, [div0, div1])
 }))
-test('hy__bind|error', with_connected_dom(async connected_dom=>{
+test('hy_op|error', with_connected_dom(async connected_dom=>{
 	const el_a:Element[] = []
-	const div0 = div({ 'hy__bind': 'no-fn' })
+	const div0 = div({ 'hy_op': 'no-fn' })
 	const fn1 = (el:Element)=>el_a.push(el)
-	const div1 = div({ 'hy__bind': 'fn1' })
+	const div1 = div({ 'hy_op': 'fn1' })
 	attach(connected_dom, div(), div0, div1, div())
-	throws(()=>hy__bind(document, { fn1 }), 'missing key: no-fn')
+	throws(()=>hy_op(document, { fn1 }), 'missing key: no-fn')
 }))
 test.run()
 function with_connected_dom(func:(dom:Element)=>void|Promise<void>) {